requirements.txt
docutils~=0.12
twine~=1.6.5
pyxcute~=0.3.0
pylint~=1.7.0
sphinx-autobuild~=0.6.0
